aprgbooleanalgebra refactoring

+++ AprgBooleanAlgebra/AprgBooleanAlgebra/src/BooleanAlgebra/Satisfiability/Utilities.cpp
@@ -19,11 +19,10 @@ unsigned int getSatisfiabilityLevel(
     unsigned int result(0);
     for(SatisfiabilityTerm const& satTerm : satTerms)
     {
-        result = max(result, satTerm.size());
+        result = max(result, static_cast<unsigned int>(satTerm.size()));
     }
     return result;
 }
-
 SatisfiabilityTerms getSatisfiabilityTerms(
         Term const& term)
 {
