+++ AprgBooleanAlgebra/AprgBooleanAlgebra/src/BooleanAlgebra/Simplification/SimplificationUtilities.cpp
@@ -23,11 +23,10 @@ namespace booleanAlgebra
 namespace Simplification
 {
 
-using Minterm = unsigned long long;
+using Minterm = uint64_t;
 using Implicant = Implicant<Minterm>;
 using Implicants = Implicants<Minterm>;
 using QuineMcCluskey = QuineMcCluskey<Minterm>;
-
 // utilties functions for this file
 namespace
 {
